+++ src/LocalizationConnector.jsx
@@ -1,0 +1,11 @@
+import React, { PropTypes } from 'react';
+
+export default (ComposedComponent) => {
+  const ConnectLocalization = props => <ComposedComponent {...props} />;
+
+  ConnectLocalization.displayName = 'ConnectLocalization';
+  ConnectLocalization.contextTypes = {
+    localize: PropTypes.func
+  };
+  return ConnectLocalization;
+};