+++ src/util/localize-formatter.js
@@ -1,10 +1,9 @@
-import get from 'lodash.get';
 import util from 'util';
 
-export default function(message, key, values) {
+export default (message, key, values) => {
   if (message) {
     return util.format(message, ...values);
   }
 
   return key;
-}
+};